Update to snapshot 13740187
