+++ settings.gradle.kts
@@ -1,4 +1,4 @@
-val snapshotVersion : String? = "13693757"
+val snapshotVersion : String? = "13740187"
 
 pluginManagement {
     repositories {
